refactor: Rename current store file
to recreate store using more standard way
